:recycle: chore(api) move certification attestation route
