--- api/src/certification/results/application/certification-attestation-route.js
@@ -1,104 +1,0 @@
-import Joi from 'joi';
-
-import { securityPreHandlers } from '../../../shared/application/security-pre-handlers.js';
-import { LOCALE } from '../../../shared/domain/constants.js';
-import { identifiersType } from '../../../shared/domain/types/identifiers-type.js';
-import { certificationController } from './certification-controller.js';
-
-const { FRENCH_SPOKEN, ENGLISH_SPOKEN } = LOCALE;
-
-const register = async function (server) {
-  const adminRoutes = [
-    {
-      method: 'GET',
-      path: '/api/admin/sessions/{sessionId}/attestations',
-      config: {
-        validate: {
-          params: Joi.object({
-            sessionId: identifiersType.sessionId,
-          }),
-        },
-        pre: [
-          {
-            method: (request, h) =>
-              securityPreHandlers.hasAtLeastOneAccessOf([
-                securityPreHandlers.checkAdminMemberHasRoleSuperAdmin,
-                securityPreHandlers.checkAdminMemberHasRoleCertif,
-                securityPreHandlers.checkAdminMemberHasRoleSupport,
-              ])(request, h),
-            assign: 'hasAuthorizationToAccessAdminScope',
-          },
-        ],
-        handler: certificationController.getCertificationPDFAttestationsForSession,
-        plugins: {
-          'hapi-swagger': {
-            produces: ['application/pdf'],
-          },
-        },
-        notes: [
-          '- **Route accessible par un user Admin**\n' +
-            "- Récupération des certifications d'une session au format PDF" +
-            ' via un id de session et un user id',
-        ],
-        tags: ['api', 'certifications', 'PDF'],
-      },
-    },
-  ];
-
-  server.route([
-    ...adminRoutes,
-    {
-      method: 'GET',
-      path: '/api/attestation/{certificationCourseId}',
-      config: {
-        validate: {
-          params: Joi.object({
-            certificationCourseId: identifiersType.certificationCourseId,
-          }),
-          query: Joi.object({
-            isFrenchDomainExtension: Joi.boolean().required(),
-            lang: Joi.string().valid(FRENCH_SPOKEN, ENGLISH_SPOKEN),
-          }),
-        },
-        handler: certificationController.getPDFAttestation,
-        notes: [
-          '- **Route accessible par un user authentifié**\n' +
-            '- Récupération des informations d’une certification au format PDF' +
-            ' via un id de certification et un user id',
-        ],
-        tags: ['api', 'certifications', 'PDF'],
-      },
-    },
-    {
-      method: 'GET',
-      path: '/api/organizations/{organizationId}/certification-attestations',
-      config: {
-        pre: [
-          {
-            method: securityPreHandlers.checkUserIsAdminInSCOOrganizationManagingStudents,
-            assign: 'belongsToOrganizationManagingStudents',
-          },
-        ],
-        validate: {
-          params: Joi.object({
-            organizationId: identifiersType.organizationId,
-          }),
-          query: Joi.object({
-            division: Joi.string().required(),
-            isFrenchDomainExtension: Joi.boolean().required(),
-            lang: Joi.string().valid(FRENCH_SPOKEN, ENGLISH_SPOKEN),
-          }),
-        },
-        handler: certificationController.downloadCertificationAttestationsForDivision,
-        tags: ['api', 'organizations'],
-        notes: [
-          'Cette route est restreinte aux utilisateurs authentifiés',
-          "Elle retourne les certificats par classe liées à l'organisation sous forme de fichier PDF.",
-        ],
-      },
-    },
-  ]);
-};
-
-const name = 'certification-attestation-api';
-export { name, register };
